+++ packages/sdk/src/types.ts
@@ -38,7 +38,7 @@ export interface ClusterSlotType {
 export interface ShipLike {
     coordinates: ServerContract.Types.coordinates
     hullmass?: UInt32
-    energy?: UInt16
+    energy?: UInt32
     engines?: ServerContract.Types.movement_stats
     generator?: ServerContract.Types.energy_stats
     loader_lanes?: ServerContract.Types.loader_lane[]
